On branch ArtBranch Your branch is up to date with 'origin/ArtBranch'. Changes to be committed:     modified:   src/App.css     modified:   src/App.js     modified:   src/assets/styles/main.css     modified:   src/layouts/Footer/Footer.css     modified:   src/layouts/Footer/Footer.jsx     modified:   src/layouts/Header/Header.css     modified:   src/layouts/Header/Header.jsx     deleted:    src/pages/Main.jsx     new file:   src/pages/MainPage.jsx
