+++ src/assets/styles/main.css
@@ -1,0 +1,10 @@
+.page-content {
+    box-sizing: border-box;
+    background-color: #fffdf3;
+    min-height: 90vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    color: white;
+}
