+++ src/layouts/Footer/Footer.css
@@ -1,0 +1,4 @@
+.footer-main {
+    height: 50px;
+    background-color: cadetblue;
+}
