+++ src/layouts/Footer/Footer.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import "./Footer.css"
+
+const Footer = () => {
+  return (
+    <div className='footer-main'>Footer</div>
+  )
+}
+
+export default Footer
