+++ src/layouts/Header/Header.css
@@ -1,0 +1,4 @@
+.header-main {
+    height: 50px;
+    background-color: cadetblue;
+}
