+++ src/layouts/Header/Header.jsx
@@ -1,8 +1,9 @@
 import React from 'react'
+import "./Header.css"
 
 const Header = () => {
   return (
-    <div>Header</div>
+    <div className='header-main'>Header</div>
   )
 }
