--- src/pages/Main.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Main = () => {
-  return (
-    <div>Main</div>
-  )
-}
-
-export default Main
